MAINT: More logical helper functions arrangement
