+++ scipy/optimize/_lsq_trf.py
@@ -104,9 +104,9 @@ from scipy.sparse.linalg import LinearOperator, aslinearoperator, lsmr
 from .optimize import OptimizeResult
 from ._lsq_common import (
     step_size_to_bound, make_strictly_feasible, find_active_constraints,
-    scaling_vector, intersect_trust_region, solve_lsq_trust_region,
-    solve_trust_region_2d, update_tr_radius, minimize_quadratic_1d,
-    build_quadratic_1d, evaluate_quadratic, compute_grad, compute_jac_scaling,
+    intersect_trust_region, solve_lsq_trust_region, solve_trust_region_2d,
+    update_tr_radius, minimize_quadratic_1d, build_quadratic_1d,
+    evaluate_quadratic, compute_grad, compute_jac_scaling,
     check_termination, right_multiply, print_header, print_iteration)
 
 
@@ -122,6 +122,53 @@ def lsmr_operator(Jop, diag_root):
         return Jop.rmatvec(x1) + diag_root * x2
 
     return LinearOperator((m + n, n), matvec=matvec, rmatvec=rmatvec)
+
+
+def scaling_vector(x, g, lb, ub):
+    """Compute a scaling vector and its derivatives as described in papers
+    of Coleman and Li [1]_.
+
+    Components of a vector v are defined as follows:
+    ::
+
+               | ub[i] - x[i], if g[i] < 0 and ub[i] < np.inf
+        v[i] = | x[i] - lb[i], if g[i] > 0 and lb[i] > -np.inf
+               | 1,           otherwise
+
+    According to this definition v[i] >= 0 for all i. It differs from the
+    definition in paper [1]_ (eq. (2.2)), where the absolute value of v is
+    used. Both definitions are equivalent down the line.
+
+    Derivatives of v with respect to x take value 1, -1 or 0 depending on a
+    case.
+
+    Returns
+    -------
+    v : ndarray with shape of x
+        Scaling vector.
+    dv : ndarray with shape of x
+        Derivatives of v[i] with respect to x[i], diagonal elements of v's
+        Jacobian.
+
+    References
+    ----------
+    .. [1] Branch, M.A., T.F. Coleman, and Y. Li, "A Subspace, Interior,
+           and Conjugate Gradient Method for Large-Scale Bound-Constrained
+           Minimization Problems," SIAM Journal on Scientific Computing,
+           Vol. 21, Number 1, pp 1-23, 1999.
+    """
+    v = np.ones_like(x)
+    dv = np.zeros_like(x)
+
+    mask = (g < 0) & np.isfinite(ub)
+    v[mask] = ub[mask] - x[mask]
+    dv[mask] = -1
+
+    mask = (g > 0) & np.isfinite(lb)
+    v[mask] = x[mask] - lb[mask]
+    dv[mask] = 1
+
+    return v, dv
 
 
 def find_reflected_step(x, J_h, diag_h, g_h, p, p_h, d, Delta, l, u, theta):
@@ -214,8 +261,8 @@ def trf(fun, jac, x0, f0, J0, lb, ub, ftol, xtol, gtol, max_nfev, scaling,
     else:
         scale, scale_inv = scaling, 1 / scaling
 
-    v, jv = scaling_vector(x, g, lb, ub)
-    v[jv != 0] *= scale[jv != 0]
+    v, dv = scaling_vector(x, g, lb, ub)
+    v[dv != 0] *= scale[dv != 0]
     Delta = norm(x0 * scale / v**0.5)
     if Delta == 0:
         Delta = 1.0
@@ -241,7 +288,7 @@ def trf(fun, jac, x0, f0, J0, lb, ub, ftol, xtol, gtol, max_nfev, scaling,
         print_header()
 
     while nfev < max_nfev:
-        v, jv = scaling_vector(x, g, lb, ub)
+        v, dv = scaling_vector(x, g, lb, ub)
 
         g_norm = norm(g * v, ord=np.inf)
         if g_norm < gtol:
@@ -267,13 +314,13 @@ def trf(fun, jac, x0, f0, J0, lb, ub, ftol, xtol, gtol, max_nfev, scaling,
 
         # v is recomputed in the variables after applying `scaling`, note that
         # components which were identically 1 not affected.
-        v[jv != 0] *= scale[jv != 0]
+        v[dv != 0] *= scale[dv != 0]
 
         # Here we apply two types of scaling.
         d = v**0.5 * scale_inv
 
         # C = diag(g / scale) Jv
-        diag_h = g * jv * scale_inv
+        diag_h = g * dv * scale_inv
 
         # After all this were done, we continue normally.
